requirements.txt
# FastAPI
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
psycopg2-binary
asyncpg
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
python-multipart
requests
facebook-business
python-telegram-bot
aiofiles
apscheduler
openai

# База данных (SQLAlchemy + Async)
sqlalchemy[asyncio]==2.0.29
aiosqlite==0.20.0
alembic==1.13.1

# Аутентификация и безопасность
python-dotenv==1.0.1

# Работа с API
httpx==0.27.0

# Прочее
PyJWT==2.8.0
apscheduler==3.10.4 # Для фоновых задач (автопилот)
fastapi-cors==0.0.1
